Refactoring: Move model logic from `experimentation` to `nxontology_ml`
