+++ experimentation/export_model_predictions.py
@@ -9,11 +9,11 @@ from catboost import CatBoostClassifier, Pool
 from nxontology import NXOntology
 from sklearn.pipeline import Pipeline, make_pipeline
 
-from experimentation.model_utils import BiasedMaeMetric
 from nxontology_ml.data import get_efo_otar_slim, read_training_data
 from nxontology_ml.efo import NodeXrefFeatures
 from nxontology_ml.features import NodeInfoFeatures, PrepareNodeFeatures
-from nxontology_ml.model import MODEL_SEED, CatBoostDataFormatter
+from nxontology_ml.model.formatter import MODEL_SEED, CatBoostDataFormatter
+from nxontology_ml.model.utils import BiasedMaeMetric
 from nxontology_ml.text_embeddings.text_embeddings_transformer import (
     TextEmbeddingsTransformer,
 )
